`scenarios` - common functions for scenarios extracted from `aws_hospital_hubero` node
