+++ tiago_social_scenarios_sim/CMakeLists.txt
@@ -8,6 +8,7 @@ find_package(catkin REQUIRED COMPONENTS
 
 ## catkin specific configuration
 include_directories(
+  include
   ${catkin_INCLUDE_DIRS}
   ${hubero_ros_INCLUDE_DIRS}
 )
@@ -16,9 +17,17 @@ catkin_package(
   CATKIN_DEPENDS hubero_ros
 )
 
+add_library(scenario_utils
+  include/${PROJECT_NAME}/utils.h
+  src/utils.cpp
+)
+target_link_libraries(scenario_utils
+  ${hubero_ros_LIBRARIES}
+  ${catkin_LIBRARIES}
+)
+
 ## executables
 add_executable(aws_hospital_hubero src/aws_hospital_hubero.cpp)
 target_link_libraries(aws_hospital_hubero
-	${hubero_ros_LIBRARIES}
-	${catkin_LIBRARIES}
+  scenario_utils
 )
